+++ core/discord_client.py
@@ -1,6 +1,7 @@
+from core.notification_client import NotificationClient
 from core.requester import KibaResponse
 from core.requester import Requester
-from core.notification_client import NotificationClient
+
 
 class DiscordClient(NotificationClient):