+++ core/queues/message_queue_processor.py
@@ -12,11 +12,11 @@ from typing import TypeVar
 from core import logging
 from core.exceptions import InternalServerErrorException
 from core.exceptions import KibaException
+from core.notification_client import NotificationClient
 from core.queues.message_queue import MessageQueue
 from core.queues.model import Message
 # from core.slack_client import SlackClient
 from core.util.value_holder import RequestIdHolder
-from core.notification_client import NotificationClient
 
 
 class MessageProcessor(ABC):