+++ core/slack_client.py
@@ -1,8 +1,9 @@
 from typing import Optional
 
+from core.notification_client import NotificationClient
 from core.requester import KibaResponse
 from core.requester import Requester
-from core.notification_client import NotificationClient
+
 
 class SlackClient(NotificationClient):